rudimentary question backend
